+++ CHANGES.md
@@ -221,6 +221,14 @@ OpenSSL 4.0
 
    *Ryan Hooper*
 
+ * Streamlined the workaround for the connection callback function of type
+   `OSSL_HTTP_bio_cb_t` not having a `OSSL_HTTP_REQ_CTX *` parameter,
+   by adding `OSSL_HTTP_REQ_CTX_proxy_connect()` called on SSL/TLS connect.
+
+   This work was sponsored by Siemens AG.
+
+   *David von Oheimb*
+
  * Fixed CRLs with invalid ASN1_TIME in invalidityDate extensions,
    where verification incorrectly succeeded. Enforced proper
    handling of ASN1_TIME validation results so that any CRL